backend/requirements.txt
# Web framework
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-bcrypt==1.0.1

# Database
psycopg2-binary==2.9.9

# AI
openai>=1.50.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Optional: VirusTotal integration
requests==2.31.0
